[chore] move InnerHits to it's own file
